+++ src/pages/Feed/index.tsx
@@ -1,5 +1,5 @@
 import CommonHeader from "@/components/common/Header";
-import Navbar from "@/components/feature/feed/Navbar"
+import Navbar from "@/components/feature/feed/section/Navbar"
 const FeedPage = () =>{
 
   return(
